service/requirements.txt
Werkzeug>=0.14
gunicorn>=19.5.0
Flask>=0.12.3
requests>=2.23.0
python-dateutil>=2.8.1
nltk>=3.4.5
vaderSentiment